requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
django-multiselectfield==0.1.8
django-picklefield==2.0
django-widget-tweaks==1.4.5
idna==2.8
Pillow==6.1.0
python-barcode==0.10.0
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
